fixes file upload and cleans some code
