+++ main/core/Manager/CurlManager.php
@@ -24,6 +24,8 @@ class CurlManager
             curl_setopt($ch, $option, $value);
         }
 
+        curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
+
         switch ($type) {
             case 'POST':
                 $this->setPostCurl($ch, $payload);
