+++ plugin/claco-form/Resources/styles/resources/claco-form.less
@@ -140,6 +140,7 @@
     }
 
     .entry-actions {
+        margin-left: auto;
         padding-left: 10px;
         border-left: 1px solid #ccc;
